refactor: appointments as module
