+++ studio_app/rout_handlers/__init__.py
@@ -1,5 +1,6 @@
 __all__ = [
     'account',
     'book',
-    'all_appointments'
+    'all_appointments',
+    'appointments'
 ]
